pyproject.toml
[project]
name = "aether-drones"
version = "0.1.0"
description = "Aether: Drone Orchestration Platform"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.20.0",
    "pytest>=8.0",
    "pytest-asyncio",
    "boto3",
    "awsiotsdk",
    "awscrt",
    "paho-mqtt", # For tools/mock_iot_events
    "pymavlink",
    "datamodel-code-generator>=0.25.0",
    "python-dotenv",
    "ruff", # Added Linter
    "aether-common",
]

[tool.uv.workspace]
members = [
    "aether/orchestrator", 
    "aether/common",
]

[tool.uv.sources]
aether-common = { workspace = true }

[tool.pytest.ini_options]
pythonpath = [
    "aether/common/src",
    "aether/orchestrator/src",
    "aether/cloud-bridge/src"
]
testpaths = [
    "aether/common/tests",
    "aether/orchestrator/tests",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["aether_common", "src"]
